add save function

+++ js/view.js
@@ -1,23 +1,28 @@
 class View {
-
-
-
-
     /**
      * отображает в DOM'e текущее состояние приложения, навешивает необходимые обработчики на интерфейс
      */
     init(controller) {
         this.controller = controller;
         
-        let buttonAdd =  document.querySelector('#save'),
+        let buttonSave =  document.querySelector('#save'),
+            buttonAdd =  document.querySelector('#add'),
             buttonEdit =  document.querySelector('#edit'),
             buttonDelete =  document.querySelector('#delete');
 
             // notes =  this.controller.getNotes();
 
+        buttonSave.addEventListener('click', () => this.save());
         buttonAdd.addEventListener('click', () => this.add());
-        buttonEdit.addEventListener('click', (e) => this.edit(e));
-        buttonDelete.addEventListener('click', () => this.delete());
+
+        if (buttonEdit) {
+            buttonEdit.addEventListener('click', (e) => this.edit(e));
+        }
+
+        if (buttonDelete) {
+            buttonDelete.addEventListener('click', () => this.delete());
+        }
+
 
         // notes
         //     .then(response => this.render(response))
@@ -66,11 +71,12 @@ class View {
     /**
      * Добавление клиента
      */
-    add(id=null) {
+    save() {
         let values = document.querySelectorAll('#basicModal input[type="text"], #basicModal input[type="email"]'),
             customer = {},
             error = false,
-            errorClass = document.querySelector('.has-error');
+            errorClass = document.querySelector('.has-error'),
+            id = $('#basicModal').data().id;
 
         if (errorClass) {
             errorClass.classList.remove('has-error');
@@ -85,15 +91,14 @@ class View {
             customer[values[i].id] = values[i].value;
         }
 
+        if (id) {
+            customer['id'] = id;
+        }
+
         if (error) {
             return false;
         } else {
-            if (id) {
-                this.controller.edit(customer, id);
-            } else {
-                this.controller.add(customer);
-            }
-
+            this.controller.save(customer);
             $('#basicModal').modal('hide');
         }
 
@@ -111,16 +116,42 @@ class View {
         // model.add(noteValue);
     }
 
+
+    add() {
+
+
+
+        $('#basicModal').data('id', '').modal('show');
+
+
+
+    }
+
     /**
      * Редактирование клиента
      */
     edit(e) {
-        let id = e.currentTarget.parentNode.parentNode.getAttribute('data-id');
+        let tr = $(e.currentTarget).closest('tr'),
+            id = tr.find('.id').text(),
+            name = tr.find('.name').text(),
+            email = tr.find('.email').text(),
+            telephone = tr.find('.telephone').text(),
+            address = tr.find('.address').text(),
+            street = tr.find('.street').text(),
+            city = tr.find('.city').text(),
+            state = tr.find('.state').text(),
+            zip = tr.find('.zip').text();
 
-        this.add(id);
+        $('#name').val(name);
+        $('#email').val(email);
+        $('#telephone').val(telephone);
+        $('#address').val(address);
+        $('#street').val(street);
+        $('#city').val(city);
+        $('#state').val(state);
+        $('#zip').val(zip);
 
-        // $('#basicModal').modal('show');
-
+       $('#basicModal').attr('data-id', id).modal('show');
     }
 
     /**
